add code to serve a TCP/IP socket using the RSS protocol
The basic principle is to spawn a thread in parallel to the read data
thread.  This thread will setup a socket server, accept inbound
connections and serve the protocol as described by:

http://cygnusa.blogspot.com/2015/07/how-to-talk-to-radio-sky-spectrograph.html

The thread will read from a circular buffer provided by the m_fftAvgFIFO.
This FIFO is written from the GUI by the function
LMLL_Testing_SetFFTSpectra(), and respects the control over averaging that
is provided on the spectrum.

Control over the enabling/disabling of the server, the IP:PORT and the
number of channels sent to RSS are accessed from the Setup Output menu.
The settings are also written to the RASDR.cfg file and remembered.

- Its quite a lot of spaghetti to eat.  It may be fragile in fringe cases.
- There is an *insane* amount of copying being done.  Ugh.  It isnt my
  fault!  Seriously...
- I notice that there might be enough structure to "do it right" instead of the
  way its currently being done (like not processing all the FFT data,
  etc.)  Perhaps a subsequent implementation will do a better job, but I'm
  really more interested in breaking down the components into a set of
  processes.